requirements.txt
fastapi
transformers[sentencepiece]
torch
torchaudio
python-multipart
uvicorn
ffmpeg-python 